requirements.txt
python-dotenv
pandas
numpy
pymongo
pymongo[srv]==3.11
certifi
scikit-learn==1.3.2
dill
pyaml
scipy
#-e .